docker-machine v0.16.2-gitlab.26-t2d.6 to address
https://go.dev/doc/devel/release#go1.21.minor

go1.21.9 (released 2024-04-03) includes a security fix to the net/http
package, as well as bug fixes to the linker, and the go/types and net/http
packages.

go1.21.11 (released 2024-06-04) includes security fixes to the archive/zip
and net/netip packages, as well as bug fixes to the compiler, the go command,
the runtime, and the os package.